Function image shifting
