+++ dialogs.py
@@ -5,7 +5,7 @@ from PySide6.QtCore import Qt
 from PySide6.QtGui import QPixmap, QIcon
 from PySide6.QtWidgets import (
     QDialog, QFormLayout, QSpinBox, QPushButton, QHBoxLayout,
-    QVBoxLayout, QLabel, QWidget
+    QVBoxLayout, QLabel, QWidget, QRadioButton, QDialogButtonBox
 )
 from helper import resource_path
 
@@ -93,3 +93,42 @@ class AboutDialog(QDialog):
         layout.addWidget(info_label)
         layout.addStretch()
         layout.addLayout(button_layout)
+
+class MultipleChoiceDialog(QDialog):
+    def __init__(self, title, question, options, parent=None):
+        super().__init__(parent)
+        self.setWindowTitle(title)
+        self.question = question
+        self.options = options
+        self.selected_option = None
+        self._create_option_selector()
+
+    def _create_option_selector(self):
+        layout = QVBoxLayout()
+        question_label = QLabel(self.question)
+        self.option_group = []
+
+        layout.addWidget(question_label)
+
+        for option in self.options:
+            radio_button = QRadioButton(option)
+            radio_button.toggled.connect(self._update_selected_option)
+            layout.addWidget(radio_button)
+            self.option_group.append(radio_button)
+
+        button_box = QDialogButtonBox(QDialogButtonBox.StandardButton.Ok | QDialogButtonBox.StandardButton.Cancel)
+        button_box.accepted.connect(self.accept)
+        button_box.rejected.connect(self.reject)
+        layout.addWidget(button_box)
+        self.setLayout(layout)
+
+    def _update_selected_option(self):
+        self.selected_option = None
+
+        for radio_button in self.option_group:
+            if radio_button.isChecked():
+                self.selected_option = radio_button.text()
+                break
+
+    def get_selected_option(self):
+        return self.selected_option
